Refactor migration into upmigration service.
