+++ property-services/src/main/java/org/egov/pt/service/PropertyService.java
@@ -1,61 +1,43 @@
 package org.egov.pt.service;
 
-import java.io.*;
-import java.math.BigDecimal;
-import java.sql.Timestamp;
-import java.text.DecimalFormat;
-import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Function;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
 
-import com.jayway.jsonpath.JsonPath;
-import lombok.extern.slf4j.Slf4j;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.Sets;
+
 import org.egov.common.contract.request.RequestInfo;
-import org.egov.common.contract.request.Role;
-import org.egov.mdms.model.MasterDetail;
-import org.egov.mdms.model.MdmsCriteria;
-import org.egov.mdms.model.MdmsCriteriaReq;
-import org.egov.mdms.model.ModuleDetail;
 import org.egov.pt.config.PropertyConfiguration;
 import org.egov.pt.models.OwnerInfo;
 import org.egov.pt.models.Property;
 import org.egov.pt.models.PropertyCriteria;
 import org.egov.pt.models.enums.CreationReason;
 import org.egov.pt.models.enums.Status;
-import org.egov.pt.models.excel.*;
-import org.egov.pt.models.user.User;
 import org.egov.pt.models.user.UserDetailResponse;
 import org.egov.pt.models.user.UserSearchRequest;
 import org.egov.pt.models.workflow.State;
 import org.egov.pt.producer.Producer;
-import org.egov.pt.repository.*;
-import org.egov.pt.repository.builder.PropertyQueryBuilder;
-import org.egov.pt.repository.rowmapper.LegacyExcelRowMapper;
+import org.egov.pt.repository.PropertyRepository;
 import org.egov.pt.util.PTConstants;
 import org.egov.pt.util.PropertyUtil;
 import org.egov.pt.validator.PropertyValidator;
 import org.egov.pt.web.contracts.PropertyRequest;
-import org.egov.pt.web.controllers.PropertyController;
 import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.Sets;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @Slf4j
 public class PropertyService {
-
-	@Autowired
-	private PropertyUtil propertyutil;
-
-	@Autowired
-	private BoundaryService boundaryService;
 
 	@Autowired
 	private Producer producer;
@@ -86,25 +68,6 @@ public class PropertyService {
 
 	@Autowired
 	private CalculationService calculatorService;
-
-	@Autowired
-	private ExcelService excelService;
-
-	@Autowired
-	private OwnerExcelRepository ownerExcelRepository;
-	@Autowired
-	private PropertyExcelRepository propertyExcelRepository;
-
-	@Autowired
-	private UnitExcelRepository unitExcelRepository;
-	@Autowired
-	private LegacyExcelRowMapper legacyExcelRowMapper;
-	@Autowired
-	private AddressExcelRepository addressExcelRepository;
-	@Autowired
-	private PropertyPaymentExcelRepository propertyPaymentExcelRepository;
-	@Autowired
-	private ServiceRequestRepository restRepo;
 
 	/**
 	 * Enriches the Request and pushes to the Queue
@@ -385,261 +348,4 @@ public class PropertyService {
 		util.enrichOwner(userDetailResponse, properties, false);
 		return properties;
 	}
-
-	private MdmsCriteriaReq prepareMdMsRequest(String tenantId, String moduleName, List<String> names, String filter,
-			RequestInfo requestInfo) {
-
-		List<MasterDetail> masterDetails = new ArrayList<>();
-
-		names.forEach(name -> {
-			masterDetails.add(MasterDetail.builder().name(name).filter(filter).build());
-		});
-
-		ModuleDetail moduleDetail = ModuleDetail.builder().moduleName(moduleName).masterDetails(masterDetails).build();
-		List<ModuleDetail> moduleDetails = new ArrayList<>();
-		moduleDetails.add(moduleDetail);
-		MdmsCriteria mdmsCriteria = MdmsCriteria.builder().tenantId(tenantId).moduleDetails(moduleDetails).build();
-		return MdmsCriteriaReq.builder().requestInfo(requestInfo).mdmsCriteria(mdmsCriteria).build();
-	}
-
-	public void importProperties(InputStream file, InputStream matchedFile, Long skip, Long limit) throws Exception {
-		final Map<String, String> localityMap = new HashMap<String, String>();
-		localityMap.put("Adarsh Nagar_08", "MOR275");
-		localityMap.put("Adarsh_Nagar(08)", "MOR275");
-		localityMap.put("Ashok_Nagar", "MOR222");
-		localityMap.put("Gopalpur", "MOR130");
-		localityMap.put("Gopalpur_03", "MOR130");
-		localityMap.put("Gyani_Wali_Basti", "MOR273");
-		localityMap.put("Kumar_Kunj", "MOR029");
-		localityMap.put("Majholi", "MOR132");
-
-		BufferedReader br = new BufferedReader(new InputStreamReader(matchedFile));
-		String line = "";
-		Map<String, String> matched = new HashMap<>();
-		while ((line = br.readLine()) != null) {
-			String[] values = line.split(",");
-			matched.put(values[0], values[1]);
-		}
-
-		AtomicInteger numOfSuccess = new AtomicInteger();
-		AtomicInteger numOfErrors = new AtomicInteger();
-		excelService.read(file, skip, limit, (RowExcel row) -> {
-			LegacyRow legacyRow = null;
-			try {
-				legacyRow = legacyExcelRowMapper.map(row);
-				String tenantId = "up." + legacyRow.getULBName().toLowerCase();
-
-				RequestInfo userCreateRequestInfo = RequestInfo.builder().action("_create").apiId("Rainmaker").did("1")
-						.key("").msgId("20170310130900|en_IN").ver(".01").build();
-
-				User userRequest = new User();
-				userRequest.setActive(true);
-				userRequest.setMobileNumber(
-						( legacyRow.getMobile() != null && legacyRow.getMobile() != "" && (new BigDecimal(legacyRow.getMobile()).longValue()!=0)) ? legacyRow.getMobile()
-								: convertPTINToMobileNumber("5" + legacyRow.getPTIN()));
-				userRequest.setUserName(UUID.randomUUID().toString());
-				userRequest.setPassword("123456");
-				userRequest.setFatherOrHusbandName(legacyRow.getFHName() != null && legacyRow.getFHName().length() > 100
-						? legacyRow.getFHName().substring(0, 99)
-						: legacyRow.getFHName());
-				userRequest.setType("CITIZEN");
-				userRequest.setTenantId(tenantId);
-				userRequest.setPermanentCity(tenantId);
-				userRequest.setPermanentAddress(legacyRow.getAddress());
-				userRequest.setRoles(Arrays.asList(Role.builder().code("CITIZEN").build()));
-				String name = legacyRow.getOwnerName() != null && legacyRow.getOwnerName() != ""
-						? legacyRow.getOwnerName()
-						: "Owner of " + legacyRow.getPTIN();
-				// Invalid name. Only alphabets and special characters -, ',`, .
-				name = name.replaceAll("[\\$\"'<>?\\\\~`!@#$%^()+={}\\[\\]*,.:;“”‘’]*", "");
-				name = name.length() > 100 ? name.substring(0, 99) : name;
-				// System.out.println("Santised Name: " + sanitisedString);
-
-				userRequest.setName(name);
-				userRequest.setCorrespondenceCity(tenantId);
-				userRequest.setCorrespondenceAddress(legacyRow.getAddress());
-				UserDetailResponse userDetailResponse = userService.createUser(userCreateRequestInfo, userRequest);
-				User user = userDetailResponse.getUser().get(0);
-
-				Map<String, String> tokenResponse = (Map<String, String>) userService.getToken(user.getUserName(),
-						"123456", tenantId, user.getType());
-				String token = (String) tokenResponse.get("access_token");
-
-				RequestInfo requestinfo = RequestInfo
-						.builder().authToken(token).action("token").apiId("Rainmaker").did("1").key("")
-						.msgId("20170310130900|en_IN").ver(".01").userInfo(org.egov.common.contract.request.User
-								.builder().type(user.getType()).tenantId("up").userName(user.getUserName()).build())
-						.build();
-
-				String localityCode = localityMap.get(legacyRow.getLocality());
-				if (localityCode == null) {
-					log.warn("Empty locality code for the property {}", legacyRow.getLocality());
-				}
-
-				String pId = propertyutil.getIdList(requestinfo, tenantId, config.getPropertyIdGenName(),
-						config.getPropertyIdGenFormat(), 1).get(0);
-				String ackNo = propertyutil
-						.getIdList(requestinfo, tenantId, config.getAckIdGenName(), config.getAckIdGenFormat(), 1)
-						.get(0);
-				// String ackNo = propertyutil.getIdList(requestInfo, tenantId,
-				// config.getAckIdGenName(), config.getAckIdGenFormat(), 1).get(0);
-				org.egov.pt.models.excel.Property property = new org.egov.pt.models.excel.Property();
-				property.setId(UUID.randomUUID().toString());
-				property.setPropertyid(pId);
-				property.setTenantid(tenantId);
-				property.setAccountid(user.getUuid());
-				property.setStatus("APPROVED");
-				property.setAcknowldgementnumber(ackNo);
-				property.setPropertytype("BUILTUP");
-				property.setOwnershipcategory("INDIVIDUAL.SINGLEOWNER");
-				if (matched.containsKey(legacyRow.getPropertyTypeClassification())) {
-					property.setUsagecategory(matched.get(legacyRow.getPropertyTypeClassification()));
-				} else {
-					property.setUsagecategory("OTHERS");
-				}
-				// property.setCreationreason()
-				property.setNooffloors(1L);
-				property.setLandarea(BigDecimal
-						.valueOf(Double.valueOf(legacyRow.getPlotArea() != null ? legacyRow.getPlotArea() : "0")));
-				// property.setSuperbuiltuparea()
-				// property.setLinkedproperties()
-				property.setOldpropertyid(legacyRow.getPTIN());
-				property.setSource("DATA_MIGRATION");
-				property.setChannel("MIGRATION");
-				property.setConstructionyear(legacyRow.getConstructionYear());
-				property.setCreatedby(user.getUuid());
-				property.setLastmodifiedby(user.getUuid());
-				property.setCreatedtime(new Date().getTime());
-				property.setLastmodifiedtime(new Date().getTime());
-				propertyExcelRepository.save(property);
-
-				Owner owner = new Owner();
-				owner.setOwnerinfouuid(UUID.randomUUID().toString());
-				owner.setStatus(Status.ACTIVE.toString());
-				owner.setTenantid(tenantId);
-				owner.setPropertyid(property.getId());
-				owner.setUserid(user.getUuid());
-				owner.setOwnertype("NONE");
-				owner.setRelationship("FATHER");
-				owner.setCreatedby(user.getUuid());
-				owner.setLastmodifiedby(user.getUuid());
-				owner.setCreatedtime(new Date().getTime());
-				owner.setLastmodifiedtime(new Date().getTime());
-				ownerExcelRepository.save(owner);
-
-				Unit unit = new Unit();
-				unit.setId(UUID.randomUUID().toString());
-				unit.setTenantid(tenantId);
-				unit.setPropertyid(property.getId());
-				unit.setFloorno(1L);
-				if (matched.containsKey(legacyRow.getPropertyTypeClassification())) {
-					unit.setUnittype(matched.get(legacyRow.getPropertyTypeClassification()));
-				} else {
-					unit.setUnittype("OTHERS");
-				}
-
-				if (matched.containsKey(legacyRow.getPropertyTypeClassification())) {
-					unit.setUsagecategory(matched.get(legacyRow.getPropertyTypeClassification()));
-				} else {
-					unit.setUsagecategory("OTHERS");
-				}
-				unit.setOccupancytype("SELFOCCUPIED");
-				unit.setOccupancydate(0L);
-				unit.setCarpetarea(BigDecimal.valueOf(
-						Double.valueOf(legacyRow.getTotalCarpetArea() != null ? legacyRow.getTotalCarpetArea() : "0")));
-
-				// unit.setBuiltuparea(BigDecimal builtuparea)
-				// unit.setPlintharea(BigDecimal plintharea)
-				// unit.setSuperbuiltuparea(BigDecimal superbuiltuparea)
-				unit.setArv(BigDecimal
-						.valueOf(Double.parseDouble(legacyRow.getRCARV() != null ? legacyRow.getRCARV() : "0")));
-				unit.setConstructiontype("PUCCA");
-				// unit.setConstructiondate(Long constructiondate)
-				// unit.setDimensions(String dimensions)
-				unit.setActive(true);
-				unit.setCreatedby(user.getUuid());
-				unit.setLastmodifiedby(user.getUuid());
-				unit.setCreatedtime(new Date().getTime());
-				unit.setLastmodifiedtime(new Date().getTime());
-				unitExcelRepository.save(unit);
-
-				Address address = new Address();
-				address.setTenantid(tenantId);
-				address.setId(UUID.randomUUID().toString());
-				address.setPropertyid(property.getId());
-				address.setDoorno(legacyRow.getHouseNo());
-				// address.setPlotno(String plotno)
-				// address.setBuildingname(String buildingname)
-				address.setStreet(legacyRow.getAddress());
-				// address.setLandmark(String landmark)
-				address.setCity(legacyRow.getULBName());
-				address.setPincode("123456");
-				address.setLocality(localityCode);
-				// address.setLocality(legacyRow.getLocality() != null? legacyRow.getLocality():
-				// "OTHERS");
-				address.setDistrict(legacyRow.getULBName());
-				// address.setRegion(String region)
-				address.setState("Uttar Pradesh");
-				address.setCountry("India");
-				// address.setLatitude(BigDecimal latitude)
-				// address.setLongitude(BigDecimal longitude)
-				address.setCreatedby(user.getUuid());
-				address.setLastmodifiedby(user.getUuid());
-				address.setCreatedtime(new Date().getTime());
-				address.setLastmodifiedtime(new Date().getTime());
-				address.setTaxward(legacyRow.getTaxWard());
-				address.setWardname(legacyRow.getWardName());
-				address.setWardno(legacyRow.getWardNo());
-				address.setZone(legacyRow.getZone());
-				addressExcelRepository.save(address);
-				// address.setAdditionaldetails(String additionaldetails)
-
-				PropertyPayment payment = new PropertyPayment();
-				payment.setId(UUID.randomUUID().toString());
-				payment.setPropertyid(property.getId());
-				payment.setFinancialyear(legacyRow.getFinancialYear());
-				payment.setArrearhousetax(BigDecimal.valueOf(
-						Double.valueOf(legacyRow.getArrearHouseTax() != null ? legacyRow.getArrearHouseTax() : "0")));
-				payment.setArrearwatertax(BigDecimal.valueOf(
-						Double.valueOf(legacyRow.getArrearWaterTax() != null ? legacyRow.getArrearWaterTax() : "0")));
-				payment.setArrearsewertax(BigDecimal.valueOf(
-						Double.valueOf(legacyRow.getArrearSewerTax() != null ? legacyRow.getArrearSewerTax() : "0")));
-				payment.setHousetax(BigDecimal
-						.valueOf(Double.valueOf(legacyRow.getHouseTax() != null ? legacyRow.getHouseTax() : "0")));
-				payment.setWatertax(BigDecimal
-						.valueOf(Double.valueOf(legacyRow.getWaterTax() != null ? legacyRow.getWaterTax() : "0")));
-				payment.setSewertax(BigDecimal
-						.valueOf(Double.valueOf(legacyRow.getSewerTax() != null ? legacyRow.getSewerTax() : "0")));
-				payment.setSurcharehousetax(BigDecimal.valueOf(Double
-						.valueOf(legacyRow.getSurchareHouseTax() != null ? legacyRow.getSurchareHouseTax() : "0")));
-				payment.setSurcharewatertax(BigDecimal.valueOf(Double
-						.valueOf(legacyRow.getSurchareWaterTax() != null ? legacyRow.getSurchareWaterTax() : "0")));
-				payment.setSurcharesewertax(BigDecimal.valueOf(Double
-						.valueOf(legacyRow.getSurchareSewerTax() != null ? legacyRow.getSurchareSewerTax() : "0")));
-				payment.setBillgeneratedtotal(BigDecimal.valueOf(Double
-						.valueOf(legacyRow.getBillGeneratedTotal() != null ? legacyRow.getBillGeneratedTotal() : "0")));
-				payment.setTotalpaidamount(BigDecimal.valueOf(
-						Double.valueOf(legacyRow.getTotalPaidAmount() != null ? legacyRow.getTotalPaidAmount() : "0")));
-
-				payment.setLastpaymentdate(legacyRow.getLastPaymentDate());
-				propertyPaymentExcelRepository.save(payment);
-				numOfSuccess.getAndIncrement();
-			} catch (Exception e) {
-				numOfErrors.getAndIncrement();
-				System.out
-						.println("Row[" + row.getRowIndex() + "] - [" + legacyRow.toString() + "] -" + e.getMessage());
-			}
-			return true;
-		});
-		System.out.println("Import Completed - Success=" + numOfSuccess + " Errors=" + numOfErrors);
-	}
-
-	private String convertPTINToMobileNumber(String ptin) {
-		String curPtin = ptin;
-		while (curPtin.length() < 9) {
-			curPtin = "0" + curPtin;
-		}
-		return "5" + curPtin;
-	}
-	
 }
