+++ property-services/src/main/java/org/egov/pt/web/controllers/PropertyController.java
@@ -1,6 +1,5 @@
 package org.egov.pt.web.controllers;
 
-import java.io.IOException;
 import java.io.InputStream;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -15,22 +14,26 @@ import org.egov.pt.models.PropertyCriteria;
 import org.egov.pt.models.oldProperty.OldPropertyCriteria;
 import org.egov.pt.service.MigrationService;
 import org.egov.pt.service.PropertyService;
+import org.egov.pt.service.UPMigrationService;
 import org.egov.pt.util.ResponseInfoFactory;
 import org.egov.pt.validator.PropertyValidator;
 import org.egov.pt.web.contracts.PropertyRequest;
 import org.egov.pt.web.contracts.PropertyResponse;
 import org.egov.pt.web.contracts.RequestInfoWrapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
 @RequestMapping("/property")
 public class PropertyController {
-
 
 	final ClassLoader loader = PropertyController.class.getClassLoader();
 
@@ -38,35 +41,35 @@ public class PropertyController {
 	private PropertyService propertyService;
 
 	@Autowired
+	private UPMigrationService upMigrationService;
+
+	@Autowired
 	private ResponseInfoFactory responseInfoFactory;
 
 	@Autowired
 	private MigrationService migrationService;
-	
+
 	@Autowired
-    private PropertyValidator propertyValidator;
+	private PropertyValidator propertyValidator;
 
 	@PostMapping("/_create")
 	public ResponseEntity<PropertyResponse> create(@Valid @RequestBody PropertyRequest propertyRequest) {
 
 		Property property = propertyService.createProperty(propertyRequest);
-		ResponseInfo resInfo = responseInfoFactory.createResponseInfoFromRequestInfo(propertyRequest.getRequestInfo(), true);
-		PropertyResponse response = PropertyResponse.builder()
-				.properties(Arrays.asList(property))
-				.responseInfo(resInfo)
+		ResponseInfo resInfo = responseInfoFactory.createResponseInfoFromRequestInfo(propertyRequest.getRequestInfo(),
+				true);
+		PropertyResponse response = PropertyResponse.builder().properties(Arrays.asList(property)).responseInfo(resInfo)
 				.build();
 		return new ResponseEntity<>(response, HttpStatus.CREATED);
 	}
 
-	
 	@PostMapping("/_update")
 	public ResponseEntity<PropertyResponse> update(@Valid @RequestBody PropertyRequest propertyRequest) {
-		
+
 		Property property = propertyService.updateProperty(propertyRequest);
-		ResponseInfo resInfo = responseInfoFactory.createResponseInfoFromRequestInfo(propertyRequest.getRequestInfo(), true);
-		PropertyResponse response = PropertyResponse.builder()
-				.properties(Arrays.asList(property))
-				.responseInfo(resInfo)
+		ResponseInfo resInfo = responseInfoFactory.createResponseInfoFromRequestInfo(propertyRequest.getRequestInfo(),
+				true);
+		PropertyResponse response = PropertyResponse.builder().properties(Arrays.asList(property)).responseInfo(resInfo)
 				.build();
 		return new ResponseEntity<>(response, HttpStatus.OK);
 	}
@@ -74,14 +77,14 @@ public class PropertyController {
 	@PostMapping("/_search")
 	public ResponseEntity<PropertyResponse> search(@Valid @RequestBody RequestInfoWrapper requestInfoWrapper,
 			@Valid @ModelAttribute PropertyCriteria propertyCriteria) {
-		
-		if(propertyCriteria.getLocality()!=null && !propertyCriteria.getLocality().isEmpty())
-		{
+
+		if (propertyCriteria.getLocality() != null && !propertyCriteria.getLocality().isEmpty()) {
 			propertyCriteria.setStreet(propertyCriteria.getLocality());
 		}
-		
+
 		propertyValidator.validatePropertyCriteria(propertyCriteria, requestInfoWrapper.getRequestInfo());
-		List<Property> properties = propertyService.searchProperty(propertyCriteria,requestInfoWrapper.getRequestInfo());
+		List<Property> properties = propertyService.searchProperty(propertyCriteria,
+				requestInfoWrapper.getRequestInfo());
 		PropertyResponse response = PropertyResponse.builder().properties(properties).responseInfo(
 				responseInfoFactory.createResponseInfoFromRequestInfo(requestInfoWrapper.getRequestInfo(), true))
 				.build();
@@ -90,49 +93,55 @@ public class PropertyController {
 
 	@PostMapping("/_migration")
 	public ResponseEntity<?> propertyMigration(@Valid @RequestBody RequestInfoWrapper requestInfoWrapper,
-											   @Valid @ModelAttribute OldPropertyCriteria propertyCriteria) {
+			@Valid @ModelAttribute OldPropertyCriteria propertyCriteria) {
 		long startTime = System.nanoTime();
 		Map<String, String> resultMap = null;
 		Map<String, String> errorMap = new HashMap<>();
 
-		resultMap = migrationService.initiateProcess(requestInfoWrapper,propertyCriteria,errorMap);
+		resultMap = migrationService.initiateProcess(requestInfoWrapper, propertyCriteria, errorMap);
 
 		long endtime = System.nanoTime();
 		long elapsetime = endtime - startTime;
-		System.out.println("Elapsed time--->"+elapsetime);
+		System.out.println("Elapsed time--->" + elapsetime);
 
 		return new ResponseEntity<>(resultMap, HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/_plainsearch", method = RequestMethod.POST)
 	public ResponseEntity<PropertyResponse> plainsearch(@Valid @RequestBody RequestInfoWrapper requestInfoWrapper,
-														@Valid @ModelAttribute PropertyCriteria propertyCriteria) {
-		List<Property> properties = propertyService.searchPropertyPlainSearch(propertyCriteria, requestInfoWrapper.getRequestInfo());
+			@Valid @ModelAttribute PropertyCriteria propertyCriteria) {
+		List<Property> properties = propertyService.searchPropertyPlainSearch(propertyCriteria,
+				requestInfoWrapper.getRequestInfo());
 		PropertyResponse response = PropertyResponse.builder().properties(properties).responseInfo(
 				responseInfoFactory.createResponseInfoFromRequestInfo(requestInfoWrapper.getRequestInfo(), true))
 				.build();
 		return new ResponseEntity<>(response, HttpStatus.OK);
 	}
-//	@RequestMapping(value = "/_cancel", method = RequestMethod.POST)
-//	public ResponseEntity<PropertyResponse> cancel(@Valid @RequestBody RequestInfoWrapper requestInfoWrapper,
-//												   @Valid @ModelAttribute PropertyCancelCriteria propertyCancelCriteria) {
-//		
-//		List<Property> properties = propertyService.cancelProperty(propertyCancelCriteria,requestInfoWrapper.getRequestInfo());
-//		PropertyResponse response = PropertyResponse.builder().properties(properties).responseInfo(
-//				responseInfoFactory.createResponseInfoFromRequestInfo(requestInfoWrapper.getRequestInfo(), true))
-//				.build();
-//		return new ResponseEntity<>(response, HttpStatus.OK);
-//	}
+	// @RequestMapping(value = "/_cancel", method = RequestMethod.POST)
+	// public ResponseEntity<PropertyResponse> cancel(@Valid @RequestBody
+	// RequestInfoWrapper requestInfoWrapper,
+	// @Valid @ModelAttribute PropertyCancelCriteria propertyCancelCriteria) {
+	//
+	// List<Property> properties =
+	// propertyService.cancelProperty(propertyCancelCriteria,requestInfoWrapper.getRequestInfo());
+	// PropertyResponse response =
+	// PropertyResponse.builder().properties(properties).responseInfo(
+	// responseInfoFactory.createResponseInfoFromRequestInfo(requestInfoWrapper.getRequestInfo(),
+	// true))
+	// .build();
+	// return new ResponseEntity<>(response, HttpStatus.OK);
+	// }
 
 	@PostMapping("/_import")
-	public ResponseEntity<?> propertyImport(@RequestParam(required = false) Long limit, @RequestParam(required = false, defaultValue = "1") Long skip) throws Exception {
+	public ResponseEntity<?> propertyImport(@RequestParam(required = false) Long limit,
+			@RequestParam(required = false, defaultValue = "1") Long skip) throws Exception {
 		long startTime = System.nanoTime();
 		final InputStream excelFile = loader.getResourceAsStream("legacy-bareilly.xlsx");
 		final InputStream matchedFile = loader.getResourceAsStream("matched.csv");
-		propertyService.importProperties(excelFile, matchedFile, skip, limit);
+		upMigrationService.importProperties(excelFile, matchedFile, skip, limit);
 		long endtime = System.nanoTime();
 		long elapsetime = endtime - startTime;
-		System.out.println("Elapsed time--->"+elapsetime);
+		System.out.println("Elapsed time--->" + elapsetime);
 
 		return new ResponseEntity<>(true, HttpStatus.OK);
 	}
